Abstract custom metrics exporter API server (#19)
