+++ .pylintrc
@@ -88,7 +88,7 @@ persistent=yes
 
 # Minimum Python version to use for version dependent checks. Will default to
 # the version used to run pylint.
-py-version=3.10
+py-version=3.12
 
 # Discover python modules and packages in the file system subtree.
 recursive=no
@@ -426,7 +426,8 @@ disable=raw-checker-failed,
         use-symbolic-message-instead,
         C0114,
         C0115,
-        C0116
+        C0116,
+        R0903
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
