+++ redis-queue-hpa/Chart.yaml
@@ -1,6 +1,11 @@
 ---
 apiVersion: v2
 name: suihei
-version: 0.0.1
+description: Redis Queue HPA
+type: application
+# chart version
+version: 0.0.3
+# application version
+appVersion: "1.0.0"
 
 dependencies: []
