+++ redis-queue-hpa/Makefile
@@ -1,9 +1,13 @@
 default: docker
 
+# rqmw is redis-queue-monitor-worker
 .PHONY: docker
 docker:
-	@img="zannen/redis-queue-monitor-worker" && \
-	ver=$$(python -c 'import yaml; print(yaml.safe_load(open("values.yaml","r"))["rqmwVersion"])') && \
-	cd docker && \
-	echo "Building $$img:$$ver" && \
-	docker build -t "$$img:$$ver" redis-queue-monitor-worker/
+	@for i in metricsexporter rqmw ; do \
+		img="zannen/$$i" && \
+		ver=$$(python -c 'import yaml; print(yaml.safe_load(open("values.yaml","r"))["'"$$i"'Version"])') && \
+		cd docker && \
+		echo "Building $$img:$$ver" && \
+		docker build -t "$$img:$$ver" "$$i/" && \
+		cd .. ; \
+	done
