+++ redis-queue-hpa/docker/metricsexporter/requirements.txt
@@ -1,0 +1,4 @@
+Flask==3.1.0
+gunicorn==23.0.0
+redis==5.2.1
+rq==2.1.0
