+++ redis-queue-hpa/templates/apiserver-custom-metrics.yaml
@@ -28,7 +28,7 @@ spec:
     spec:
       containers:
         - name: custom-metrics-apiserver
-          image: "zannen/redis-queue-monitor-worker:{{.Values.rqmwVersion}}"
+          image: "zannen/metricsexporter:{{.Values.metricsexporterVersion}}"
           imagePullPolicy: Never
           command:
             - gunicorn
@@ -61,11 +61,13 @@ metadata:
   namespace: {{.Release.Namespace}}
 spec:
   ports:
-    - port: 443
-      protocol: TCP
-      targetPort: 6443
+    - protocol: TCP
+      targetPort: 6443  # in container
+      port: 443         # across cluster on ClusterIP
+      # nodePort        # from outside cluster
   selector:
     app: custom-metrics-apiserver
+  type: NodePort
 
 # {{if .Values.customMetrics.apiService.deploy}}
 ---
