+++ redis-queue-hpa/templates/redis-server.yaml
@@ -65,8 +65,9 @@ metadata:
 spec:
   type: ClusterIP
   ports:
-    - port: 6379
-      protocol: TCP
-      targetPort: 6379
+    - protocol: TCP
+      targetPort: 6379  # in container
+      port: 6379        # across cluster on ClusterIP
+      # nodePort        # from outside cluster
   selector:
     app: redis-server
